CER::aCL(): Cleanup
git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@15703 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/gui/rulers/ControllerEventsRuler.h
@@ -77,7 +77,10 @@ public:
      * Does not return anything yet, as whether it should or not has yet to be
      * determined.
      */
-    virtual void addControlLine(float x1, float y1, float x2, float y2, bool eraseExistingContollers);
+    virtual void addControlLine(
+            double x1, double y1,
+            double x2, double y2,
+            bool eraseExistingContollers);
 
     /** Draw a rubber band indicating the controller line that will be drawn if
      * the user clicks another event into existence while moving the pencil
